web/requirements.txt
Flask
Flask_sqlalchemy
Flask_marshmallow
Flask_migrate
flask_jwt_extended
Jinja2
psycopg2-binary
pyyaml